rename MetricFactory to metrics and unify to single utility 'Metrics'
