+++ services/src/main/java/io/scalecube/services/ServiceDispatcherFactory.java
@@ -1,6 +1,6 @@
 package io.scalecube.services;
 
-import io.scalecube.services.metrics.MetricFactory;
+import io.scalecube.services.metrics.Metrics;
 import io.scalecube.services.routing.Router;
 import io.scalecube.services.routing.RouterFactory;
 
@@ -14,7 +14,7 @@ public class ServiceDispatcherFactory {
     this.routerFactory = new RouterFactory(serviceRegistry);
   }
 
-  public ServiceCall createDispatcher(Class<? extends Router> routerType, Duration timeout, MetricFactory metrics) {
+  public ServiceCall createDispatcher(Class<? extends Router> routerType, Duration timeout, Metrics metrics) {
     return new ServiceCall(routerFactory.getRouter(routerType), timeout, metrics);
   }
